Purge
These files are removed due to the following reasons.
I. The previouse architecture was monolith with the frontend and the backend
integrated together which add load to the server. The new architecture splits the
frontend and the backend server into two with django and django rest framework in
the backend talking over HTTPS API with react application on a separate server. Thus
These deleted files are irelevant.
II. Since the backend does not handle serving the pages over HTTP the old views are
irelevant and will be replaced with API views serving JSON response to the react server